requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
executor==23.2
fasteners==0.19
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
nest-asyncio==1.6.0
property-manager==3.0
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.1.0
python-telegram-bot==21.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
verboselogs==1.7
yarl==1.9.4
